rag_retrieval/requirements.txt
openai>=1.0.0
faiss-cpu>=1.7.4
numpy>=1.24.0
streamlit>=1.28.0
fastapi>=0.109.0
uvicorn[standard]>=0.27.0
pydantic>=2.5.0
